+++ hal/disk_hal.h
@@ -6,8 +6,8 @@
 
 struct disk_hal;
 
-typedef bool(*disk_hal_read_func)(struct disk_hal *params, uint32_t sector, void *buf);
-typedef bool(*disk_hal_write_func)(struct disk_hal *params, uint32_t sector, void *buf);
+typedef bool(*disk_hal_read_func)(struct disk_hal *params, uint64_t sector, void *buf);
+typedef bool(*disk_hal_write_func)(struct disk_hal *params, uint64_t sector, void *buf);
 
 struct disk_hal_ops {
   disk_hal_read_func read_op;
@@ -21,8 +21,8 @@ struct disk_hal {
 };
 
 struct disk_hal *disk_hal_init();
-bool disk_hal_read(struct disk_hal *disk, uint32_t sector, void *buf);
-bool disk_hal_write(struct disk_hal *disk, uint32_t sector, void *buf);
+bool disk_hal_read(struct disk_hal *disk, uint64_t sector, void *buf);
+bool disk_hal_write(struct disk_hal *disk, uint64_t sector, void *buf);
 bool disk_hal_destory(struct disk_hal *disk);
